Answer the review: the score key takes the board's own types
impl Display let a caller smuggle any formatting into the persisted
leaderboard identity, which undermined the pin. score_key now takes
the same concrete u64 pair build_board takes.

+++ faces/cli/src/main.rs
@@ -5493,7 +5493,10 @@ fn munch_with_input(
             .map(|n| n - 1)
             .collect();
         let outcome = numinous_core::grade_munch(&board, &bites);
-        post_score(&numinous_core::munch_score_key(seed, round), outcome.score);
+        post_score(
+            &numinous_core::munch_score_key(seed, round as u64),
+            outcome.score,
+        );
         if numinous_core::munch_clean_win(&outcome) {
             journey.win();
             println!(
